1blog

+++ Blog_Limiun/src/main/java/com/example/demo/mapper/UserMapper.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserMapper {
-    User SelectUser(long id);
+    User SelectUser(int id);
     User GetUserByAccount(String account);
-    int InsertUser(long id,String username,String account,String password,int sex ,int age,int phone,int account_state,String special_sign,long money,String extendMap);
+    int InsertUser(int id,String username,String account,String password,int sex ,int age,int phone,int account_state,String special_sign,long money,String extendMap);
 }
